fix: BodyType 쿼리 시 경계값 처리

See merge request s11-bigdata-recom-sub1/S11P21C106!201

+++ project/back/health/src/main/java/com/ssafy/health/domain/body/BodyPredict/service/BodyPredictReadService.java
@@ -107,8 +107,8 @@ public class BodyPredictReadService {
 
     private Long findBodyTypeByBmi(Float height, Float weight, Gender gender) {
 
-        Float bmi = weight / (height * height) * 10000;
-        Optional<BodyType> bodyType = bodyTypeRepository.findByBmiAndGender(bmi, gender);
+        float bmi = weight / (height * height) * 10000;
+        Optional<BodyType> bodyType = bodyTypeRepository.findByBmiAndGender((int) bmi, gender);
 
         if (bodyType.isPresent()) {
             return bodyType.get().getId();
